Completely refactor internal structure
